# FileOperationInAzure/FileOperationInAzure/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FileOperationInAzure",
    "version": "v1"
  },
  "paths": {
    "/Home": {
      "get": {
        "tags": [
          "Home"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Home/UploadFile": {
      "post": {
        "tags": [
          "Home"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "ContentType",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ContentDisposition",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Headers",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Length",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "formData",
            "name": "Name",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "FileName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Home/DeleteFile/{filename}": {
      "delete": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    },
    "/WeatherForecast/TestString": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      }
    }
  }
}